Added deity check on delete.

+++ class/Controller/Admin/Organization.php
@@ -58,6 +58,9 @@ class Organization extends SubController
 
     protected function delete(Request $request)
     {
+        if (!\Current_User::isDeity()) {
+            throw new PrivilegeMissing();
+        }
         $organization = self::load();
         OrganizationFactory::delete($organization);
         return ['success' => 1];
